Move files to match other repositories
