+++ bin/extract-pg-schema
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+// eslint-disable-next-line unicorn/prefer-module
+require("../build/cli.js");
