+++ package.json
@@ -14,7 +14,7 @@
   "main": "build/index.js",
   "types": "build/index.d.ts",
   "bin": {
-    "extract-pg-schema": "build/bin/extract-pg-schema.js"
+    "extract-pg-schema": "bin/extract-pg-schema"
   },
   "files": [
     "build/"
